requirements.txt
bs4
google
grpcio
requests
